backend/layers/shared/requirements.txt
pg8000==1.30.2
sqlalchemy==2.0.23
pyjwt==2.8.0
boto3==1.34.0
